Warning fix.

+++ fastlib/trunk/contrib/dongryel/thesis_research/core/parallel/reference_tree_walker.h
@@ -269,7 +269,8 @@ class ReferenceTreeWalker {
 
           // Pop the next item in the list, if we should continue.
           should_terminate =
-            ((world.size() == 1 && essential_reference_subtrees_[world.rank()].size() >= max_hashed_subtrees_to_queue) ||
+            ((world.size() == 1 &&
+              static_cast<int>(essential_reference_subtrees_[world.rank()].size()) >= max_hashed_subtrees_to_queue) ||
              (world.size() > 1 && static_cast<int>(hashed_essential_reference_subtrees_to_send->size()) >= max_hashed_subtrees_to_queue)  || trace_.empty());
 
           if(! should_terminate) {
@@ -284,7 +285,8 @@ class ReferenceTreeWalker {
 
         // The termination condition.
         should_terminate =
-          ((world.size() == 1 && essential_reference_subtrees_[world.rank()].size() >= max_hashed_subtrees_to_queue) ||
+          ((world.size() == 1 &&
+            static_cast<int>(essential_reference_subtrees_[world.rank()].size()) >= max_hashed_subtrees_to_queue) ||
            (world.size() > 1 && static_cast<int>(hashed_essential_reference_subtrees_to_send->size()) >= max_hashed_subtrees_to_queue) ||
            trace_.empty());
         if(should_terminate) {
